backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
pydantic-settings>=1.0.0
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
joblib>=1.2.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
